* Aditya dev2 (#62)

* added ListPackages

* finished with time out

* dkfrif

* done

* hopefully fix

* Aditya dev2 (#64)

* added ListPackages

* finished with time out

* dkfrif

* done

* hopefully fix

* Aditya dev2 (#66)

* added ListPackages

* finished with time out

* dkfrif

* done

* hopefully fix

* finished upload pt1

* Dockerfile to deploy to Cloud Run

* Remove Dockerfile

* Get PORT from environment

* Aditya dev2 (#67)

* added ListPackages

* finished with time out

* dkfrif

* done

* hopefully fix

* finished upload pt1

* fixed upload and finished rate

* Aditya dev2 (#68)

* added ListPackages

* finished with time out

* dkfrif

* done

* hopefully fix

* finished upload pt1

* fixed upload and finished rate

* fixed and tested authentication token

* fixed and tested authentication token

---------


---------

+++ src/handlers/package.go
@@ -4,13 +4,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"net/http"
-  "os"
+	"os"
 	"os/exec"
 	"path"
 	"regexp"
-  "time"
+	"time"
+	"strconv"
 
 	"tomr/models"
 	"tomr/src/db"
@@ -21,7 +21,8 @@ import (
 )
 
 const pkgDirPath = "src/metrics/temp" // temp directory to store packages
-const auth_success = "ABC"
+// const auth_success = "ABC"
+const auth_success = "bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiZWNlMzA4NjFkZWZhdWx0YWRtaW51c2VyIiwicGFzc3dvcmQiOiJjb3JyZWN0aG9yc2ViYXR0ZXJ5c3RhcGxlMTIzKCFfXytAKiooQeKAmeKAnWA7RFJPUCBUQUJMRSBwYWNrYWdlczsifQ.TSGs6VJMFx5NV2RoHrhEP_FK8nv4Wlzc4gQls2JYPC4"
 const bucket_name = "tomr"
 
 type metadata struct {
@@ -30,61 +31,152 @@ type metadata struct {
 	ID      string `json:"ID"`
 }
 
-type data struct {
-	Conent    string `json:"Content"`
-	URL       string `json:"URL"`
-	JSProgram string `json:"JSProgram"`
-}
-
 type Package struct {
-	Metadata metadata `json:"metadata"`
-	Data     data     `json:"data"`
+	Metadata metadata           `json:"metadata"`
+	Data     models.PackageData `json:"data"`
 }
 
-func CreatePackage(content string, url string, jsprogram string) {
-packageData := models.PackageData{Content: content, URL: url, JSProgram: jsprogram}
-	pkgDir := ""
-	metadata := models.PackageMetadata{}
-	//utils.PrintPackageData(packageData)
-	rating := models.PackageRating{}
-	var readMe []byte
-	// Return Error 400 if both Content and URL are set
-	if (packageData.Content != "") && (packageData.URL != "") {
-		fmt.Printf("Error 400: Content and URL cannot be both set")
-	} else if packageData.Content != "" { // Only Content is set
-		// Decode base64 string into zip
-		pkgDir = path.Join(pkgDirPath, "package.zip")
-		utils.Base64ToZip(packageData.Content, pkgDir)
-		err := utils.GetMetadataFromZip(pkgDir, &metadata, &readMe)
-		if err != nil {
-			log.Fatalf("Failed to get metadata from zip file\n")
-		}
-		metadata.ID = metadata.Name + "(" + metadata.Version + ")"
-		err = metrics.RatePackage(metadata.RepoURL, pkgDir, &rating, metadata.License, &readMe)
-	} else { // Only URL is set
-		gitUrl := utils.GetGithubUrl(url)
-		pkgDir = utils.CloneRepo(gitUrl, pkgDirPath)
-		err := metrics.RatePackage(gitUrl, pkgDir, &rating, "", nil)
-		if err != nil {
-			log.Fatalf("Failed to rate package at URL: %s\n", url)
-		}
-		// Check if package meets criteria for ingestion
-		utils.GetPackageMetadata(pkgDir, &metadata)
-		err = utils.ZipDirectory(pkgDir, pkgDir+".zip")
-		if err != nil {
-			log.Fatal(err)
-		}
-		pkgDir += ".zip"
-	}
-	utils.PrintMetadata(metadata)
-	utils.PrintRating(rating)
-	fmt.Printf(pkgDir)
-	// Upload package and store data in system
-	pkg := models.PackageObject{Metadata: &metadata, Data: &packageData, Rating: &rating}
+func CreatePackage(writer http.ResponseWriter, request *http.Request) {
 
-	err := db.StorePackage(pkg, pkgDir)
-	if err != nil {
-		log.Fatal(err)
+	request.ParseForm()
+
+	var given_xAuth string
+
+	if request.Form["X-Authorization"] != nil {
+		given_xAuth = request.Form["X-Authorization"][0]
+	} else {
+		given_xAuth = request.Header["X-Authorization"][0]
+	}
+
+	if given_xAuth == auth_success {
+		var data models.PackageData
+		body, _ := ioutil.ReadAll(request.Body)
+		json.Unmarshal(body, &data)
+
+		content := data.Content
+		url := data.URL
+		jsprogram := data.JSProgram
+
+		packageData := models.PackageData{Content: content, URL: url, JSProgram: jsprogram}
+		pkgDir := ""
+		metadata := models.PackageMetadata{}
+		//utils.PrintPackageData(packageData)
+		rating := models.PackageRating{}
+		var readMe []byte
+		// Return Error 400 if both Content and URL are set
+		if (packageData.Content != "") && (packageData.URL != "") {
+			writer.WriteHeader(400)
+			writer.Write([]byte("There is missing field(s) in the PackageData/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."))
+			os.RemoveAll(pkgDirPath)
+			return
+
+		} else if packageData.Content != "" { // Only Content is set
+			// Decode base64 string into zip
+			pkgDir = path.Join(pkgDirPath, "package.zip")
+			utils.Base64ToZip(packageData.Content, pkgDir)
+			err := utils.GetMetadataFromZip(pkgDir, &metadata, &readMe)
+			if err != nil {
+				writer.WriteHeader(400)
+				writer.Write([]byte("Error no metadata in zip"))
+				// log.Fatalf("Failed to get metadata from zip file\n")
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+			metadata.ID = metadata.Name + "(" + metadata.Version + ")"
+			if db.DoesPackageExist(metadata.ID) {
+				writer.WriteHeader(409)
+				writer.Write([]byte("Package exists already."))
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+
+			err = metrics.RatePackage(metadata.RepoURL, pkgDir, &rating, metadata.License, &readMe)
+			if err != nil {
+				writer.WriteHeader(424)
+				writer.Write([]byte("Package is not uploaded due to the disqualified rating."))
+				// log.Fatalf("Failed to get metadata from zip file\n")
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+
+			writer.WriteHeader(201)
+			writer.Write([]byte("Success. Check the ID in the returned metadata for the official ID."))
+
+		} else { // Only URL is set
+			gitUrl := utils.GetGithubUrl(url)
+			pkgDir = utils.CloneRepo(gitUrl, pkgDirPath)
+			err := metrics.RatePackage(gitUrl, pkgDir, &rating, "", nil)
+			if err != nil {
+				fmt.Print(err)
+				writer.WriteHeader(424)
+				writer.Write([]byte("Package is not uploaded due to the disqualified rating."))
+				// log.Fatalf("Failed to get metadata from zip file\n")
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+			// Check if package meets criteria for ingestion
+			utils.GetPackageMetadata(pkgDir, &metadata)
+
+			if db.DoesPackageExist(metadata.ID) {
+				writer.WriteHeader(409)
+				writer.Write([]byte("Package exists already."))
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+
+			err = utils.ZipDirectory(pkgDir, pkgDir+".zip")
+			if err != nil {
+				writer.WriteHeader(400)
+				writer.Write([]byte("Unable to zip directory"))
+				// log.Fatalf("Failed to get metadata from zip file\n")
+				os.RemoveAll(pkgDirPath)
+				return
+			}
+			pkgDir += ".zip"
+
+			writer.WriteHeader(201)
+			writer.Write([]byte("Success. Check the ID in the returned metadata for the official ID.\n"))
+		}
+
+		pkg := models.PackageObject{Metadata: &metadata, Data: &packageData, Rating: &rating}
+
+		base64, error1 := utils.ZipToBase64(pkgDir)
+
+		if error1 != nil {
+			os.RemoveAll(pkgDirPath)
+			fmt.Print("could not convert to base64")
+			return
+		}
+
+		writePath := "src/db/upload.txt"
+		fileWriter, err2 := os.Create(writePath)
+		fileWriter.WriteString(base64)
+
+		if err2 != nil {
+			os.RemoveAll(writePath)
+			os.RemoveAll(pkgDirPath)
+			fmt.Print("Failed to write base64 encoding : ", err2)
+			return
+		}
+
+		return_json, err := db.StorePackage(pkg, writePath)
+		if err != nil {
+			os.RemoveAll(writePath)
+			os.RemoveAll(pkgDirPath)
+			fmt.Print("did not get return json")
+		}
+
+		var returnVal db.Return_storage
+		json.Unmarshal(return_json, &returnVal)
+
+		returnVal.Data.Content = base64
+		return_json, _ = json.MarshalIndent(returnVal, "", "  ")
+
+		writer.Write([]byte(string(return_json)))
+
+		os.RemoveAll(writePath)
+		os.RemoveAll(pkgDirPath)
+
 	}
 }
 
@@ -238,7 +330,7 @@ func RetrievePackage(writer http.ResponseWriter, request *http.Request) {
 		return_package.Metadata.Name = rs[1]
 		return_package.Metadata.Version = rs[2]
 		return_package.Metadata.ID = metadata["ID"]
-		return_package.Data.Conent = base64
+		return_package.Data.Content = base64
 		return_package.Data.URL = metadata["URL"]
 		return_package.Data.JSProgram = metadata["JSProgram"]
 
@@ -277,16 +369,26 @@ func UpdatePackage(writer http.ResponseWriter, request *http.Request) {
 
 	if given_xAuth == auth_success {
 
+		if !db.DoesPackageExist(id) {
+			writer.WriteHeader(404)
+			writer.Write([]byte("Package does not exist"))
+			return
+		}
+
 		db.DeleteFile(bucket_name, id)
 
 		var recieve_package Package
 		body, _ := ioutil.ReadAll(request.Body)
 		json.Unmarshal(body, &recieve_package)
 
-		fmt.Print(recieve_package)
+		content := recieve_package.Data.Content
+		url := recieve_package.Data.URL
+		jsprogram := recieve_package.Data.JSProgram
 
-		writer.WriteHeader(200)
-		writer.Write([]byte("Version is updated."))
+		fmt.Print(content)
+		fmt.Print(url)
+		fmt.Print(jsprogram)
+
 	} else {
 		writer.WriteHeader(400)
 		writer.Write([]byte("There is missing field(s) in the PackageID/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."))
@@ -334,7 +436,7 @@ func ListPackages(writer http.ResponseWriter, request *http.Request) {
 		given_xAuth = request.Header["X-Authorization"][0]
 	}
 
-	if given_xAuth == auth_success {		
+	if given_xAuth == auth_success {
 
 		type Pack struct {
 			Version string `json:"Version"`
@@ -351,30 +453,30 @@ func ListPackages(writer http.ResponseWriter, request *http.Request) {
 			}
 			c1 <- Packs
 		}()
-		
+
 		select {
-			case res := <-c1:
-				writer.WriteHeader(200)
+		case res := <-c1:
+			writer.WriteHeader(200)
 
-				writer.Write([]byte("[\n"))
+			writer.Write([]byte("[\n"))
 
-				for i := 0; i < len(res); i++ {
-					result := utils.Packages(res[i].Version, res[i].Name)
-					writer.Write([]byte("  "))
+			for i := 0; i < len(res); i++ {
+				result := utils.Packages(res[i].Version, res[i].Name)
+				writer.Write([]byte("  "))
 
-					writer.Write([]byte(string(result[0])))
+				writer.Write([]byte(string(result[0])))
 
-					if i != len(res)-1 {
-						writer.Write([]byte(",\n"))
-					}
+				if i != len(res)-1 {
+					writer.Write([]byte(",\n"))
 				}
+			}
 
-				writer.Write([]byte("\n]"))
-			case <-time.After(60 * time.Second):
-				writer.WriteHeader(413)
-				writer.Write([]byte("Too many packages returned."))
+			writer.Write([]byte("\n]"))
+		case <-time.After(60 * time.Second):
+			writer.WriteHeader(413)
+			writer.Write([]byte("Too many packages returned."))
 		}
-	
+
 	} else if given_xAuth == "" {
 		writer.WriteHeader(400)
 		writer.Write([]byte("There is missing field(s) in the PackageData/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."))
@@ -402,29 +504,134 @@ func RatePackage(writer http.ResponseWriter, request *http.Request) {
 
 	if given_xAuth == auth_success {
 
-		// Call go functions here
+		if !db.DoesPackageExist(id) {
+			writer.WriteHeader(404)
+			writer.Write([]byte("Package does not exist."))
+			return
+		}
 
-		writer.WriteHeader(201)
-		writer.Write([]byte("Success. Check the ID in the returned metadata for the official ID."))
-	} else if given_xAuth == "" || id == "" {
+		type ratings struct {
+			GoodPinningPractice  string
+			NetScore             string
+			PullRequest          string
+			ResponsiveMaintainer string
+			LicenseScore         string
+			RampUp               string
+			BusFactor            string
+			Correctness          string
+		}
+
+		attrs, _ := db.GetMetadata(bucket_name, id)
+		metadata := attrs.Metadata
+
+		var package_ratings ratings
+
+		package_ratings.GoodPinningPractice = metadata["GoodPinningPractice"]
+		tempFloat, err := strconv.ParseFloat(package_ratings.GoodPinningPractice, 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.NetScore = metadata["NetScore"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.NetScore, 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.PullRequest = metadata["GoodEngineeringProcess"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.PullRequest, 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.ResponsiveMaintainer = metadata["ResponsiveMaintainer"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.ResponsiveMaintainer , 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.LicenseScore = metadata["LicenseScore"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.LicenseScore , 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.RampUp = metadata["RampUp"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.RampUp , 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.BusFactor = metadata["BusFactor"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.BusFactor , 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		package_ratings.Correctness = metadata["Correctness"]
+		tempFloat, err = strconv.ParseFloat(package_ratings.Correctness , 32)
+		if err != nil || tempFloat < 0 || tempFloat > 1 {
+			writer.WriteHeader(500)
+			writer.Write([]byte("The package rating system choked on at least one of the metrics."))
+		}
+
+		writer.WriteHeader(200)
+		return_json, _ := json.MarshalIndent(package_ratings, "", "  ")
+
+		writer.Write([]byte(string(return_json)))
+		// writer.Write([]byte("Success. Check the ID in the returned metadata for the official ID."))
+	} else {
 		writer.WriteHeader(400)
 		writer.Write([]byte("There is missing field(s) in the PackageID/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."))
-	} else {
-		writer.WriteHeader(401)
 	}
 }
 
 func CreateAuthToken(writer http.ResponseWriter, request *http.Request) {
 
-	// var auth_string
+	type User_struct struct {
+		Name    string `json:"name"`
+		IsAdmin bool   `json:"isAdmin"`
+	}
 
-	// body, _ := ioutil.ReadAll(request.Body)
+	type Secret_struct struct {
+		Password string `json:"password"`
+	}
 
-	// var auth_token auth_body
-	// json.Unmarshal(body, &auth_token)
+	type Auth struct {
+		User   User_struct   `json:"user"`
+		Secret Secret_struct `json:"secret"`
+	}
 
-	// fmt.Print(auth_token)
+	var auth_struct Auth
+	body, _ := ioutil.ReadAll(request.Body)
+	// fmt.Print(body)
+	json.Unmarshal([]byte(body), &auth_struct)
 
+	// fmt.Print(auth_struct.Secret.Password)
+
+	if auth_struct == (Auth{}) || auth_struct.User == (User_struct{}) || auth_struct.Secret == (Secret_struct{}) {
+		writer.WriteHeader(400)
+		writer.Write([]byte("There is missing field(s) in the AuthenticationRequest or it is formed improperly."))
+		return
+	}
+
+	auth_token := utils.Authenticate(auth_struct.User.Name, auth_struct.Secret.Password)
+
+	if auth_token == "err" {
+		writer.WriteHeader(401)
+		writer.Write([]byte("The user or password is invalid."))
+		return
+	}
+
+	writer.WriteHeader(200)
+	writer.Write([]byte("\"bearer " + auth_token + "\""))
 }
 
 func GetPackageByName(writer http.ResponseWriter, request *http.Request) {
